refactor fields to task list column service
