+++ projects/valtimo/task/src/lib/services/index.ts
@@ -16,3 +16,4 @@
 
 export * from './task.service';
 export * from './task-list.service';
+export * from './task-list-column.service';
